requirements.txt
python-irodsclient
mysql-connector-python
